Relocate message handler mock from internal pkg
